#78 [update]enemyのスプライトをボス戦かそうではないかで変更できるようにした
